requirements.txt
numpy==1.21.2
pandas~=2.2.3
scikit-learn~=1.6.1
opencv-python==4.5.3.20210929
tensorflow==2.6.0
requests==2.26.0
torch~=2.5.1
torchvision~=0.20.1
pillow~=11.0.0
classify~=0.5.0